+++ internal/css_parser/css_parser.go
@@ -965,6 +965,38 @@ func (p *parser) convertTokensHelper(tokens []css_lexer.Token, close css_lexer.T
 	var result []css_ast.Token
 	var nextWhitespace css_ast.WhitespaceFlags
 
+	// Enable verbatim whitespace mode when the first two non-whitespace tokens
+	// are a CSS variable name followed by a colon. This is because it could be
+	// a form of CSS variable usage, and removing whitespace could potentially
+	// break this usage. For example, the following CSS is ignored by Chrome if
+	// the whitespace isn't preserved:
+	//
+	//   @supports (--foo: ) {
+	//     html { background: green; }
+	//   }
+	//
+	// Strangely whitespace removal doesn't cause the declaration to be ignored
+	// in Firefox or Safari, so there's definitely a browser bug somewhere.
+	if !opts.verbatimWhitespace {
+		for i, t := range tokens {
+			if t.Kind == css_lexer.TWhitespace {
+				continue
+			}
+			if t.Kind == css_lexer.TIdent && strings.HasPrefix(t.DecodedText(p.source.Contents), "--") {
+				for _, t := range tokens[i+1:] {
+					if t.Kind == css_lexer.TWhitespace {
+						continue
+					}
+					if t.Kind == css_lexer.TColon {
+						opts.verbatimWhitespace = true
+					}
+					break
+				}
+			}
+			break
+		}
+	}
+
 loop:
 	for len(tokens) > 0 {
 		t := tokens[0]